@bfzli | Refactor business approval and rejection logic - Removed unnecessary code related to city creation - Updated email sending logic for business approval and rejection
