+++ actions/businesses/ApproveBusiness.ts
@@ -5,11 +5,18 @@ import { Analytics, Console } from '@/helpers/logs'
 import { BusinessModel, UserModel } from '@/data/models'
 import { CurrentTimestamp } from '@/helpers/dates'
 import { ObjectId } from '@/helpers/libs/mongo'
-import { BUSINESS_STATUSES, CONTEXT_KEYS, FULL_APP_HOST, MODELS, RESEND_FROM_EMAIL } from '@/data/constants'
-import { OnBusinessApproval, OnBusinessRejection } from '@/ui/templates'
+import { OnBusinessApproval } from '@/ui/templates'
 import { SendEmail } from '@/helpers/libs/resend'
 import { Translation } from '@/helpers/generals'
 import { FormatBusinessTitle } from '@/helpers/businesses'
+
+import { 
+    BUSINESS_STATUSES, 
+    CONTEXT_KEYS, 
+    FULL_APP_HOST, 
+    MODELS,
+    RESEND_FROM_EMAIL 
+} from '@/data/constants'
 
 const ApproveBusiness = async (c: Context) => {
     try {
@@ -46,10 +53,7 @@ const ApproveBusiness = async (c: Context) => {
                     'TotalBusinesses'
                 ])
                 
-                const isApproved = business.Status === BUSINESS_STATUSES.APPROVED && !business.Mails.OnApprovalMail
-                const isRejected = business.Status === BUSINESS_STATUSES.REJECTED && !business.Mails.OnRejectionMail
-
-                if (isApproved) {
+                if (!business.Mails.OnApprovalMail) {
                     const businessTitle = FormatBusinessTitle(business?.Title)
                     const businessLink = `${FULL_APP_HOST}/${business?.Slug}`
 
@@ -75,35 +79,6 @@ const ApproveBusiness = async (c: Context) => {
                     }
             
                     else Console.Error('ApproveBusiness', 'on_welcome_email_was_not_sent')
-                }
-
-                else if (isRejected) {
-                    const businessTitle = FormatBusinessTitle(business?.Title)
-                    const businessLink = `${FULL_APP_HOST}/${business?.Slug}/edit`
-
-                    const subject = `${business?.User?.Name}, ${Translation('your-business')} "${businessTitle}" ${Translation('is-rejected')} ❌`
-
-                    const onRejectionStatus = await SendEmail({
-                        subject,
-                        from: `${Translation('app-name')} <${RESEND_FROM_EMAIL}>`,
-                        toEmail: business?.User?.Email,
-                        template: OnBusinessRejection({
-                            subject,
-                            userName: business?.User?.Name || '',
-                            businessName: businessTitle,
-                            businessLink,
-                            reasonOfRejection: business.RejectionNote || ''
-                        })
-                    })
-
-                    if (onRejectionStatus) {
-                        business.Mails.OnRejectionMail = true
-                        business.Updated_At = CurrentTimestamp()
-            
-                        await business.save()
-                    }
-            
-                    else Console.Error('ApproveBusiness', 'on_rejection_email_was_not_sent')
                 }
 
                 return await HttpResponder({
