+++ actions/businesses/RejectBusiness.ts
@@ -3,9 +3,19 @@ import type { Context } from 'hono'
 import { DecodeBody, HttpResponder } from '@/helpers/http'
 import { Console } from '@/helpers/logs'
 import { BusinessModel, UserModel } from '@/data/models'
-import { CurrentTimestamp } from '@/helpers/dates'
+import { CurrentTimestamp, TimestampPlusDays } from '@/helpers/dates'
 import { ObjectId } from '@/helpers/libs/mongo'
-import { BUSINESS_STATUSES, CONTEXT_KEYS, USER_ROLES } from '@/data/constants'
+import { FormatBusinessTitle } from '@/helpers/businesses'
+import { Translation } from '@/helpers/generals'
+import { SendEmail } from '@/helpers/libs/resend'
+import { OnBusinessRejection } from '@/ui/templates'
+
+import { 
+    BUSINESS_STATUSES, 
+    CONTEXT_KEYS, 
+    FULL_APP_HOST,
+    RESEND_FROM_EMAIL 
+} from '@/data/constants'
 
 const SubmitBusiness = async (c: Context) => {
     try {
@@ -22,6 +32,7 @@ const SubmitBusiness = async (c: Context) => {
             if (business) {
                 business.Status = BUSINESS_STATUSES.REJECTED
                 business.RejectionNote = rejectionNote
+                business.Delete_At = TimestampPlusDays(30, 'days')
                 business.Updated_At = CurrentTimestamp()
     
                 await business.save()
@@ -34,6 +45,35 @@ const SubmitBusiness = async (c: Context) => {
                         HasPendingBusinessSubmission: false
                     }
                 )
+
+                if (!business.Mails.OnRejectionMail) {
+                    const businessTitle = FormatBusinessTitle(business?.Title)
+                    const businessLink = `${FULL_APP_HOST}/${business?.Slug}/edit`
+
+                    const subject = `${business?.User?.Name}, ${Translation('your-business')} "${businessTitle}" ${Translation('is-rejected')} ❌`
+
+                    const onRejectionStatus = await SendEmail({
+                        subject,
+                        from: `${Translation('app-name')} <${RESEND_FROM_EMAIL}>`,
+                        toEmail: business?.User?.Email,
+                        template: OnBusinessRejection({
+                            subject,
+                            userName: business?.User?.Name || '',
+                            businessName: businessTitle,
+                            businessLink,
+                            reasonOfRejection: business.RejectionNote || ''
+                        })
+                    })
+
+                    if (onRejectionStatus) {
+                        business.Mails.OnRejectionMail = true
+                        business.Updated_At = CurrentTimestamp()
+            
+                        await business.save()
+                    }
+            
+                    else Console.Error('ApproveBusiness', 'on_rejection_email_was_not_sent')
+                }
     
                 return await HttpResponder({
                     c,
