--- actions/cities/index.ts
@@ -1,7 +1,0 @@
-import CreateCity from '@/actions/cities/CreateCity'
-import ListCities from '@/actions/cities/ListCities'
-
-export { 
-    CreateCity, 
-    ListCities 
-}
