--- actions/countries/index.ts
@@ -1,7 +1,0 @@
-import CreateCountry from '@/actions/countries/CreateCountry'
-import ListCountries from '@/actions/countries/ListCountries'
-
-export { 
-    CreateCountry, 
-    ListCountries 
-}
