+++ actions/crons/index.ts
@@ -1,5 +1,7 @@
 import GenerateBackup from '@/actions/crons/GenerateBackup'
+import PerformCleanup from '@/actions/crons/PerformCleanup'
 
 export {
-    GenerateBackup
+    GenerateBackup,
+    PerformCleanup
 }
