+++ data/constants/States.ts
@@ -1,7 +1,5 @@
 import { StatesType } from '@/ts'
 
 export const STATES: StatesType = {
-    CATEGORIES: null,
-    CITIES: null,
-    COUNTRIES: null
+    CATEGORIES: null
 }
