+++ data/langs/sq_AL.ts
@@ -1,20 +1,12 @@
 const sq_AL = {
     'app-name': 'Bizneset nga Peswaktle',
     'app-description': 'Një platformë për promovimin e bizneseve me pronarë muslimanë.',
-    'country-code-should-be-string': 'Kodi i vendit duhet të jetë një string.',
-    'country-code-should-be-at-2-characters': 'Kodi i vendit duhet të jetë saktësisht 2 karaktere.',
-    'city-name-min-error': 'Emri i qytetit duhet të jetë së paku 2 karaktere.',
-    'city-name-max-error': 'Emri i qytetit duhet të jetë së paku 64 karaktere.',
-    'city-selection-error': 'Ju lutem zgjidhni një qytet valid.',
     'saves-where-listed-successfully': 'Ruajtjet u listuan me sukses.',
     'boolean-error-validation': 'Vlera e dhënë nuk është e vlefshme!',
     'much': 'shumë',
-    'city-name-should-be-string': 'Emri i qytetit duhet të jetë vetëm tekst.',
     'similar-businesses-where-listed-successfully': 'Bizneset e ngjashme u listuan me sukses.',
-    'country-post-selection-error': 'Ju lutem zgjidhni një vend valid.',
     'list-of-categories-was-retrived-but-empty': 'Lista e kategorive u mor por është e zbrazët.',
     'user-is-not-authorized-to-create-a-category': 'Përdoruesi nuk është i autorizuar të krijojë një kategori.',
-    'city-was-created-successfully': 'Qyteti u krijua me sukses.',
     'something-went-wrong-while-listing-saves': 'Diçka shkoi keq gjatë listimit të ruajtjeve.',
     'business-was-not-found': 'Biznesi nuk u gjet.',
     'add-location': 'Shto vendndodhje',
@@ -57,33 +49,16 @@ const sq_AL = {
     'user-could-not-be-updated-beacuse-the-body-data-is-invalid': 'Përdoruesi nuk mund të përditësohet sepse të dhënat e trupit janë të pavlefshme.',
     'user-was-updated-successfully': 'Përdoruesi u përditësua me sukses.',
     'user-was-not-found-so-user-could-not-be-updated': 'Përdoruesi nuk u gjet kështu që përdoruesi nuk mund të përditësohet.',
-    'list-of-cities-was-retrieved-successfully': 'Lista e qyteteve u mor me sukses.',
-    'list-of-cities-was-retrived-but-empty': 'Lista e qyteteve u mor por është e zbrazët.',
-    'list-of-cities-could-not-be-retrieved': 'Lista e qyteteve nuk mund të marrët.',
-    'list-of-countries-was-retrieved-successfully': 'Lista e vendeve u mor me sukses.',
-    'list-of-countries-was-retrived-but-empty': 'Lista e vendeve u mor por është e zbrazët.',
-    'list-of-countries-could-not-be-retrieved': 'Lista e vendeve nuk mund të marrët.',
     'list-of-categories-could-not-be-retrieved': 'Lista e kategorive nuk mund të marrët.',
     'invalid-api-language': 'Gjuha e API-së është e pavlefshme.',
     'list-of-categories-was-retrieved-successfully': 'Lista e kategorive u mor me sukses.',
     'invalid-api-version': 'Versioni i API-së është i pavlefshëm.',
     'user-was-not-authed-because-of-a-problem': 'Përdoruesi nuk u autentikua për shkak të një problemi.',
     'route-was-not-found': 'Rruga per te ky funksion nuk u gjet.',
-    'city-could-not-be-created': 'Qyteti nuk mund të krijohet.',
     'category-was-created-successfully': 'Kategoria u krijua me sukses.',
     'category-could-not-be-created': 'Kategoria nuk mund të krijohet.',
-    'user-is-not-authorized-to-create-a-city': 'Përdoruesi nuk është i autorizuar të krijojë një qytet.',
-    'country-not-found-so-city-could-not-be-created': 'Vendi nuk u gjet kështu që qyteti nuk mund të krijohet.',
-    'country-was-created-successfully': 'Vendi u krijua me sukses.',
-    'country-could-not-be-created': 'Vendi nuk mund të krijohet.',
     'user-was-loged-in-successfully': 'Përdoruesi u kyç me sukses.',
-    'user-is-not-authorized-to-create-a-country': 'Përdoruesi nuk është i autorizuar të krijojë një vend.',
-    'country-was-not-created-because-of-validation-errors': 'Vendi nuk u krijua për shkak të gabimeve të validimit.',
-    'country-name-should-be-string': 'Emri i vendit duhet të jetë string!',
     'backup-was-generated-successfully': 'Kopja rezervë u gjenerua me sukses.',
-    'city-was-not-created-because-of-validation-errors': 'Qyteti nuk u krijua për shkak të gabimeve të validimit.',
-    'country-name-min-error': 'Emri i vendit duhet të jetë më i gjatë se 2 karaktere!',
-    'country-name-max-error': 'Emri i vendit duhet të jetë më i shkurtë se 64 karaktere!',
     'category-was-not-created-because-of-validation-errors': 'Kategoria nuk u krijua për shkak të gabimeve të validimit.',
     'category-not-found-so-category-could-not-be-created': 'Kategoria nuk u gjet kështu që kategoria nuk mund të krijohet.',
     'name-min-error': 'Name must be longer than 2 characters!',
