+++ data/models/index.ts
@@ -2,8 +2,6 @@ import AnalyticModel from '@/data/models/AnalyticModel'
 import BackupModel from '@/data/models/BackupModel'
 import BusinessModel from '@/data/models/BusinessModel'
 import CategoryModel from '@/data/models/CategoryModel'
-import CityModel from '@/data/models/CityModel'
-import CountryModel from '@/data/models/CountryModel'
 import LogModel from '@/data/models/LogModel'
 import UserModel from '@/data/models/UserModel'
 import SaveModel from '@/data/models/SaveModel'
@@ -13,8 +11,6 @@ export {
     BackupModel,
     BusinessModel,
     CategoryModel,
-    CityModel,
-    CountryModel,
     LogModel,
     UserModel,
     SaveModel
