+++ data/structures/index.ts
@@ -2,8 +2,6 @@ import AnalyticStructure from '@/data/structures/AnalyticStructure'
 import BackupStructure from '@/data/structures/BackupStructure'
 import BusinessStructure from '@/data/structures/BusinessStructure'
 import CategoryStructure from '@/data/structures/CategoryStructure'
-import CityStructure from '@/data/structures/CityStructure'
-import CountryStructure from '@/data/structures/CountryStructure'
 import LogStructure from '@/data/structures/LogStructure'
 import SaveStructure from '@/data/structures/SaveStructure'
 import UserStructure from '@/data/structures/UserStructure'
@@ -13,8 +11,6 @@ export {
     BackupStructure,
     BusinessStructure,
     CategoryStructure,
-    CityStructure,
-    CountryStructure,
     LogStructure,
     SaveStructure,
     UserStructure
