+++ helpers/entities/index.ts
@@ -1,9 +1,5 @@
 import GetCategories from '@/helpers/entities/GetCategories'
-import GetCities from '@/helpers/entities/GetCities'
-import GetCountries from '@/helpers/entities/GetCountries'
 
 export {
-    GetCategories,
-    GetCities,
-    GetCountries
+    GetCategories
 }
