+++ helpers/validations/index.ts
@@ -1,22 +1,12 @@
-import CityNameValidation from '@/helpers/validations/CityNameValidation'
-import CityValidation from '@/helpers/validations/CityValidation'
-import CountryNameValidation from '@/helpers/validations/CountryNameValidation'
 import CategoryNameValidation from '@/helpers/validations/CategoryNameValidation'
 import CategorySlugValidation from '@/helpers/validations/CategorySlugValidation'
-import CountryValidation from '@/helpers/validations/CountryValidation'
-import CountryCodeValidation from '@/helpers/validations/CountryCodeValidation'
 import NotificationBooleanValidation from '@/helpers/validations/NotificationBooleanValidation'
 import UserNameValidation from '@/helpers/validations/UserNameValidation'
 import UserSurnameValidation from '@/helpers/validations/UserSurnameValidation'
 
 export {
-    CityNameValidation,
-    CityValidation,
-    CountryNameValidation,
     CategoryNameValidation,
     CategorySlugValidation,
-    CountryValidation,
-    CountryCodeValidation,
     NotificationBooleanValidation,
     UserNameValidation,
     UserSurnameValidation
