+++ router/CronsRouter.ts
@@ -3,9 +3,8 @@ import type { HonoBase } from 'hono/hono-base'
 import { Hono } from 'hono'
 import { KeyMiddleware } from '@/helpers/middlewares'
 import { Connect } from '@/helpers/libs/mongo'
-import { GenerateBackup } from '@/actions/crons'
+import { GenerateBackup, PerformCleanup } from '@/actions/crons'
 import { CRONS_ROUTES } from '@/data/constants'
-import PerformCleanup from '@/actions/crons/PerformCleanup'
 
 const CronsRouter = (): HonoBase => {
     const router = new Hono()
