+++ router/index.ts
@@ -1,7 +1,5 @@
 import BusinessesRouter from '@/router/BusinessesRouter'
 import CategoriesRouter from '@/router/CategoriesRouter'
-import CitiesRouter from '@/router/CitiesRouter'
-import CountriesRouter from '@/router/CountriesRouter'
 import CronsRouter from '@/router/CronsRouter'
 import GeneralsRouter from '@/router/GeneralsRouter'
 import SavesRouter from '@/router/SavesRouter'
@@ -10,8 +8,6 @@ import UsersRouter from '@/router/UsersRouter'
 export {
     BusinessesRouter,
     CategoriesRouter,
-    CitiesRouter,
-    CountriesRouter,
     CronsRouter,
     GeneralsRouter,
     SavesRouter,
